changed from single element push to batched version

+++ queue_rs/src/fs.rs
@@ -1,12 +1,12 @@
 use std::path::Path;
 use std::{fs, io};
-pub fn dir_size(path: &String) -> io::Result<u64> {
-    fn dir_size(mut dir: fs::ReadDir) -> io::Result<u64> {
+pub fn dir_size(path: &String) -> io::Result<usize> {
+    fn dir_size(mut dir: fs::ReadDir) -> io::Result<usize> {
         dir.try_fold(0, |acc, file| {
             let file = file?;
             let size = match file.metadata()? {
                 data if data.is_dir() => dir_size(fs::read_dir(file.path())?)?,
-                data => data.len(),
+                data => data.len() as usize,
             };
             Ok(acc + size)
         })
